+++ src/ui.h
@@ -27,8 +27,8 @@ void resetUI(void);
 
 LCDBitmap* getTitleNewGameBitmap(uint8_t _i);
 
-int32_t getScrollOffset(void);
+float getScrollOffset(void);
 
-int32_t modScrollOffset(int32_t _mod);
+float modScrollOffset(float _mod);
 
-void setScrollOffset(int32_t _set);
+void setScrollOffset(float _set);
